pyproject.toml
[project]
name = "data-tester"
version = "0.0.1"
description = "Plattform for testing data and ETL logic"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pydantic",  # DTOs and data validation
    "pydantic-settings",  # for reading from ENV vars
    "polars",  # used for fixtures crunching - esp. for testcase COMPARE
    "xlsxwriter",  # required for polars to write excel
    "fastexcel",  # required for polars to read excel
    "pyarrow",  # required as backend for polars
    "pyyaml",  # read and write yaml files
    "duckdb",  # needed for demo and for e2e tests
    "fsspec",  # needed for demo, 2e2 tests and also for input finders and report storage
    "fastapi",  # HTTP backend framework
    "uvicorn[standard]",  # ASGI server for FastAPI
    "httpx",  # async HTTP client (used by client_interface and tests)
    "nicegui",  # NiceGUI frontend framework
    "pip>=26.0.1",
]

[project.optional-dependencies]
gcs = ["gcsfs"]  # needed if data is stored in GCS

[dependency-groups]
dev = [
    "ruff",  # linter / syntax checker and formatter - replaces flake8
    "pytest",  # test framework
    "pytest-timeouts", # to suppress timeouts during tests
    "pytest-asyncio",  # async test support
    "coverage",  # test coverage reports
    "pygount",  # count lines of code  # enforce import rules for hexagonal architecture
    "import-linter",
    "ty>=0.0.26",  # type checker
]


# UV CONFIG

[tool.uv]
default-groups = ["dev"]  # uv will install dev dependency group by default


# PYTEST CONFIG

[tool.pytest.ini_options]
minversion = "8.1"
testpaths = ["tests",]
setup_timeout = 5
execution_timeout = 300
teardown_timeout = 5
markers = [
    "skip_test: explicitely skip tests",
    "infrastructure: mark test as requiring infrastructure, e.g. a DataPlatform"
]
filterwarnings = [
    "ignore::UserWarning",
    "ignore::DeprecationWarning:pydantic"
]


# COVERAGE CONFIG

[tool.coverage.run]
command_line = "-m pytest -m 'not skip_test and not infrastructure'"
source = ["src/"]  # only analyze source code in src for coverage
relative_files = true  # use relative files to be picked up by github action py-cov-action/python-coverage-comment-action@v3
omit = [
    "src/infrastructure/backend/dummy/*",
    "src/infrastructure/notifier/in_memory_notifier.py"
]


[tool.coverage.report]
skip_empty = true  # empty python files like __init__.py are not included in report
show_missing = true
exclude_also = [
    "raise NotImplementedError",  # ignore abstract methods and not implemented methods
    "@(abc\\.)?abstractmethod",
]

# RUFF CONFIG

[tool.ruff]
line-length = 90
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "W", "B"]


# IMPORT-LINTER CONFIG

[tool.importlinter]
root_package = "src"

[[tool.importlinter.contracts]]
name = "1. Domain only imports ports"
type = "forbidden"
source_modules = ["src.domain"]
forbidden_modules = [
    "src.apps",
    "src.config",
    "src.domain_adapters",
    "src.drivers",
    "src.infrastructure",
]

[[tool.importlinter.contracts]]
name = "2. Drivers only import domain ports"
type = "forbidden"
source_modules = ["src.drivers"]
forbidden_modules = [
    "src.apps",
    "src.config",
    "src.domain",
    "src.domain_adapters",
    "src.infrastructure",
    "src.infrastructure_ports",
]

[[tool.importlinter.contracts]]
name = "3. Infra only imports infra ports"
type = "forbidden"
source_modules = ["src.infrastructure"]
forbidden_modules = [
    "src.apps",
    "src.config",
    "src.domain",
    "src.domain_adapters",
    "src.domain_ports",
    "src.drivers",
]

[[tool.importlinter.contracts]]
name = "4. Ports only import DTOs"
type = "forbidden"
source_modules = [
    "src.domain_ports",
    "src.infrastructure_ports"
]
forbidden_modules = [
    "src.apps",
    "src.config",
    "src.domain",
    "src.domain_adapters",
    "src.drivers",
    "src.infrastructure",
]

[[tool.importlinter.contracts]]
name = "5. Ports and config are independent"
type = "independence"
modules = [
    "src.domain_ports",
    "src.infrastructure_ports",
    "src.config",
]

[[tool.importlinter.contracts]]
name = "6. Domain hexagons are independent"
type = "independence"
modules = [
    "src.domain.*"
]

[[tool.importlinter.contracts]]
name = "7. Drivers are independent"
type = "independence"
modules = [
    "src.drivers.*"
]

[[tool.importlinter.contracts]]
name = "8. Adapters are independent"
type = "independence"
modules = [
    "src.domain_adapters.*"
]

[[tool.importlinter.contracts]]
name = "9. Adapters only import their domain objects and their interfaces"
type = "forbidden"
source_modules = ["src.domain_adapters"]
forbidden_modules = [
    "src.apps",
    "src.config",
    "src.drivers",
    "src.infrastructure",
]

[[tool.importlinter.contracts]]
name = "10. Only apps import adapters"
type = "forbidden"
source_modules = [
    "src.domain",
    "src.domain_ports",
    "src.drivers",
    "src.infrastructure",
    "src.infrastructure_ports",
    "src.dtos",
    "src.config",
]
forbidden_modules = [
    "src.domain_adapters",
]

[[tool.importlinter.contracts]]
name = "11. client_interface only imports from dtos"
type = "forbidden"
source_modules = ["src.client_interface"]
forbidden_modules = [
    "src.apps",
    "src.ui",
    "src.domain",
    "src.domain_ports",
    "src.domain_adapters",
    "src.drivers",
    "src.infrastructure",
    "src.infrastructure_ports",
    "src.config",
]

[[tool.importlinter.contracts]]
name = "12. ui only imports from client_interface and dtos"
type = "forbidden"
source_modules = ["src.ui"]
forbidden_modules = [
    "src.domain",
    "src.domain_ports",
    "src.domain_adapters",
    "src.drivers",
    "src.infrastructure",
    "src.infrastructure_ports",
    "src.apps",
    "src.config",
]

[[tool.importlinter.contracts]]
name = "13. Nothing imports from ui"
type = "forbidden"
source_modules = [
    "src.dtos",
    "src.domain",
    "src.domain_ports",
    "src.domain_adapters",
    "src.drivers",
    "src.infrastructure",
    "src.infrastructure_ports",
    "src.config",
    "src.apps",
    "src.client_interface",
]
forbidden_modules = ["src.ui"]
